requirements.txt
appdirs==1.4.4
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
decorator==5.1.1
dill==0.3.6
distlib==0.3.1
djangorestframework==3.14.0
Django==4.1.9
djongo==1.3.6
exceptiongroup==1.1.0
executing==1.2.0
filelock==3.0.12
Flask==2.3.2
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
numpy==1.24.3
packaging==23.0
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
plotly==5.14.1
pluggy==1.0.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.14.0
pylint==2.16.2
pymongo==3.12.3
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.2.4
stack-data==0.6.2
tenacity==8.2.2
tomli==2.0.1
tomlkit==0.11.6
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vboxapi==1.0
virtualenv==20.0.30
virtualenvwrapper-win==1.2.6
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
distlib==0.3.6
executing==1.2.0
filelock==3.10.7
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.14.0
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
virtualenv==20.21.0
wcwidth==0.2.6
